requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
concurrent-log-handler==0.9.25
cron-descriptor==1.4.5
cryptography==43.0.1
decorator==5.1.1
Deprecated==1.2.14
Django==5.1.1
django-allauth==65.0.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
django-simple-history==3.7.0
django-timezone-field==7.0
executing==2.1.0
flake8==7.1.1
flower==2.0.1
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-pubsub==2.26.1
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
gspread==6.1.3
gspread-dataframe==4.0.0
gspread-formatting==1.2.0
httplib2==0.22.0
humanize==4.11.0
idna==3.10
importlib_metadata==8.4.0
iniconfig==2.0.0
ipython==8.28.0
jedi==0.19.1
kombu==5.4.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mysqlclient==2.2.4
numpy==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pluggy==1.5.0
portalocker==2.10.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pytest==8.3.3
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rsa==4.9
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
wrapt==1.16.0
zipp==3.20.2
